\name{ttest2stan-package}
\alias{ttest2stan-package}
\alias{ttest2stan}
\docType{package}
\title{
\packageTitle{ttest2stan}
}
\description{
\packageDescription{ttest2stan}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ttest2stan}
\packageIndices{ttest2stan}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{ttest2stan}

Maintainer: \packageMaintainer{ttest2stan}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from ~~
~~ file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
